Move log types to a different file
